backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.9.2
python-dotenv==1.0.1
openpyxl
google-genai
anthropic
Pillow
openai
PyMuPDF
python-multipart
watchdog
slowapi==0.1.9
pyhwp
six
